Fix - refactored code and files for tooltip functionality
